ext/http11: modify the C Ragel parser to upper-snake-case headers in-place
This is based on Zed's suggestion and helps take complexity out
of the hand-written C code, allowing memcpy() to be used instead.

Zed Shaw wrote in <20080303044659.5a550c19.zedshaw@zedshaw.com>:
> * Also, now that I think about it, if you don't care that the original
> string is modified in place then you can just have ragel do all of this
> as it goes.  Simply modify the parser to have it do this transform on
> the header chars using the existing pointer.  That'd probably be
> alright since people don't usually keep the input headers around when
> using the mongrel parser.

I don't have a working Java runtime, so I've only made the bare
minimum modification to the http11_parser.java.rl file which
allows Ragel to still work with it.  All the other Java parts
are untouched and whatever upper-casing routine was used before
continues to be used now.


git-svn-id: svn://rubyforge.org/var/svn/mongrel/trunk@990 19e92222-5c0b-0410-8929-a290d50e31e9